Reorganize code search repo management files
Splitting these up and making sure we have better separation of concerns between the repo tracker and manager
