+++ src/extension/workspaceChunkSearch/vscode-node/workspaceIndexingStatus.ts
@@ -7,7 +7,7 @@ import { t } from '@vscode/l10n';
 import * as vscode from 'vscode';
 import { ILogService } from '../../../platform/log/common/logService';
 import { ICodeSearchAuthenticationService } from '../../../platform/remoteCodeSearch/node/codeSearchRepoAuth';
-import { RepoStatus, ResolvedRepoEntry } from '../../../platform/remoteCodeSearch/node/codeSearchRepoTracker';
+import { RepoStatus, ResolvedRepoEntry } from '../../../platform/workspaceChunkSearch/node/codeSearch/repoManager';
 import { LocalEmbeddingsIndexStatus } from '../../../platform/workspaceChunkSearch/node/embeddingsChunkSearch';
 import { IWorkspaceChunkSearchService, WorkspaceIndexState } from '../../../platform/workspaceChunkSearch/node/workspaceChunkSearchService';
 import { coalesce } from '../../../util/vs/base/common/arrays';
