+++ src/platform/workspaceChunkSearch/node/codeSearch/codeSearchChunkSearch.ts
@@ -4,43 +4,43 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as l10n from '@vscode/l10n';
-import { shouldInclude } from '../../../util/common/glob';
-import { Result } from '../../../util/common/result';
-import { TelemetryCorrelationId } from '../../../util/common/telemetryCorrelationId';
-import { coalesce } from '../../../util/vs/base/common/arrays';
-import { raceCancellationError, raceTimeout, timeout } from '../../../util/vs/base/common/async';
-import { CancellationToken, CancellationTokenSource } from '../../../util/vs/base/common/cancellation';
-import { isCancellationError } from '../../../util/vs/base/common/errors';
-import { Emitter, Event } from '../../../util/vs/base/common/event';
-import { Iterable } from '../../../util/vs/base/common/iterator';
-import { Lazy } from '../../../util/vs/base/common/lazy';
-import { Disposable } from '../../../util/vs/base/common/lifecycle';
-import { isEqual, isEqualOrParent } from '../../../util/vs/base/common/resources';
-import { StopWatch } from '../../../util/vs/base/common/stopwatch';
-import { URI } from '../../../util/vs/base/common/uri';
-import { IInstantiationService } from '../../../util/vs/platform/instantiation/common/instantiation';
-import { ChatResponseWarningPart } from '../../../vscodeTypes';
-import { IAuthenticationChatUpgradeService } from '../../authentication/common/authenticationUpgrade';
-import { FileChunkAndScore } from '../../chunking/common/chunk';
-import { ComputeBatchInfo } from '../../chunking/common/chunkingEndpointClient';
-import { ConfigKey, IConfigurationService } from '../../configuration/common/configurationService';
-import { EmbeddingType } from '../../embeddings/common/embeddingsComputer';
-import { RelativePattern } from '../../filesystem/common/fileTypes';
-import { GithubRepoId } from '../../git/common/gitService';
-import { logExecTime, LogExecTime, measureExecTime } from '../../log/common/logExecTime';
-import { ILogService } from '../../log/common/logService';
-import { IAdoCodeSearchService } from '../../remoteCodeSearch/common/adoCodeSearchService';
-import { IGithubCodeSearchService } from '../../remoteCodeSearch/common/githubCodeSearchService';
-import { CodeSearchResult } from '../../remoteCodeSearch/common/remoteCodeSearch';
-import { BuildIndexTriggerReason, CodeSearchRepoManager, IndexedRepoEntry, RepoEntry, RepoStatus, ResolvedRepoEntry, TriggerIndexingError } from '../../remoteCodeSearch/node/codeSearchRepoTracker';
-import { IExperimentationService } from '../../telemetry/common/nullExperimentationService';
-import { ITelemetryService } from '../../telemetry/common/telemetry';
-import { IWorkspaceService } from '../../workspace/common/workspaceService';
-import { IWorkspaceChunkSearchStrategy, StrategySearchResult, StrategySearchSizing, WorkspaceChunkQueryWithEmbeddings, WorkspaceChunkSearchOptions, WorkspaceChunkSearchStrategyId } from '../common/workspaceChunkSearch';
-import { CodeSearchWorkspaceDiffTracker } from './codeSearchWorkspaceDiff';
-import { EmbeddingsChunkSearch } from './embeddingsChunkSearch';
-import { TfIdfWithSemanticChunkSearch } from './tfidfWithSemanticChunkSearch';
-import { IWorkspaceFileIndex } from './workspaceFileIndex';
+import { shouldInclude } from '../../../../util/common/glob';
+import { Result } from '../../../../util/common/result';
+import { TelemetryCorrelationId } from '../../../../util/common/telemetryCorrelationId';
+import { coalesce } from '../../../../util/vs/base/common/arrays';
+import { raceCancellationError, raceTimeout, timeout } from '../../../../util/vs/base/common/async';
+import { CancellationToken, CancellationTokenSource } from '../../../../util/vs/base/common/cancellation';
+import { isCancellationError } from '../../../../util/vs/base/common/errors';
+import { Emitter, Event } from '../../../../util/vs/base/common/event';
+import { Iterable } from '../../../../util/vs/base/common/iterator';
+import { Lazy } from '../../../../util/vs/base/common/lazy';
+import { Disposable } from '../../../../util/vs/base/common/lifecycle';
+import { isEqual, isEqualOrParent } from '../../../../util/vs/base/common/resources';
+import { StopWatch } from '../../../../util/vs/base/common/stopwatch';
+import { URI } from '../../../../util/vs/base/common/uri';
+import { IInstantiationService } from '../../../../util/vs/platform/instantiation/common/instantiation';
+import { ChatResponseWarningPart } from '../../../../vscodeTypes';
+import { IAuthenticationChatUpgradeService } from '../../../authentication/common/authenticationUpgrade';
+import { FileChunkAndScore } from '../../../chunking/common/chunk';
+import { ComputeBatchInfo } from '../../../chunking/common/chunkingEndpointClient';
+import { ConfigKey, IConfigurationService } from '../../../configuration/common/configurationService';
+import { EmbeddingType } from '../../../embeddings/common/embeddingsComputer';
+import { RelativePattern } from '../../../filesystem/common/fileTypes';
+import { GithubRepoId } from '../../../git/common/gitService';
+import { logExecTime, LogExecTime, measureExecTime } from '../../../log/common/logExecTime';
+import { ILogService } from '../../../log/common/logService';
+import { IAdoCodeSearchService } from '../../../remoteCodeSearch/common/adoCodeSearchService';
+import { IGithubCodeSearchService } from '../../../remoteCodeSearch/common/githubCodeSearchService';
+import { CodeSearchResult } from '../../../remoteCodeSearch/common/remoteCodeSearch';
+import { IExperimentationService } from '../../../telemetry/common/nullExperimentationService';
+import { ITelemetryService } from '../../../telemetry/common/telemetry';
+import { IWorkspaceService } from '../../../workspace/common/workspaceService';
+import { IWorkspaceChunkSearchStrategy, StrategySearchResult, StrategySearchSizing, WorkspaceChunkQueryWithEmbeddings, WorkspaceChunkSearchOptions, WorkspaceChunkSearchStrategyId } from '../../common/workspaceChunkSearch';
+import { EmbeddingsChunkSearch } from '../embeddingsChunkSearch';
+import { TfIdfWithSemanticChunkSearch } from '../tfidfWithSemanticChunkSearch';
+import { IWorkspaceFileIndex } from '../workspaceFileIndex';
+import { BuildIndexTriggerReason, CodeSearchRepoManager, IndexedRepoEntry, RepoEntry, RepoStatus, ResolvedRepoEntry, TriggerIndexingError } from './repoManager';
+import { CodeSearchWorkspaceDiffTracker } from './workspaceDiff';
 
 export interface CodeSearchDiffState {
 	readonly totalFileCount: number;
