+++ src/platform/workspaceChunkSearch/node/codeSearch/repoManager.ts
@@ -3,34 +3,26 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 import * as l10n from '@vscode/l10n';
-import { exec } from 'child_process';
-import { promisify } from 'util';
-import { Result } from '../../../util/common/result';
-import { TelemetryCorrelationId } from '../../../util/common/telemetryCorrelationId';
-import { coalesce } from '../../../util/vs/base/common/arrays';
-import { CancelablePromise, createCancelablePromise, DeferredPromise, IntervalTimer, raceCancellationError, raceTimeout } from '../../../util/vs/base/common/async';
-import { CancellationToken, CancellationTokenSource } from '../../../util/vs/base/common/cancellation';
-import { isCancellationError } from '../../../util/vs/base/common/errors';
-import { Emitter, Event } from '../../../util/vs/base/common/event';
-import { Disposable } from '../../../util/vs/base/common/lifecycle';
-import { ResourceMap } from '../../../util/vs/base/common/map';
-import { Schemas } from '../../../util/vs/base/common/network';
-import { URI } from '../../../util/vs/base/common/uri';
-import { IInstantiationService } from '../../../util/vs/platform/instantiation/common/instantiation';
-import { IAuthenticationService } from '../../authentication/common/authentication';
-import { IGitExtensionService } from '../../git/common/gitExtensionService';
-import { AdoRepoId, getGithubRepoIdFromFetchUrl, getOrderedRemoteUrlsFromContext, getOrderedRepoInfosFromContext, GithubRepoId, IGitService, parseRemoteUrl, RepoContext, ResolvedRepoRemoteInfo } from '../../git/common/gitService';
-import { Change } from '../../git/vscode/git';
-import { logExecTime, LogExecTime } from '../../log/common/logExecTime';
-import { ILogService } from '../../log/common/logService';
-import { isGitHubRemoteRepository } from '../../remoteRepositories/common/utils';
-import { ISimulationTestContext } from '../../simulationTestContext/common/simulationTestContext';
-import { ITelemetryService } from '../../telemetry/common/telemetry';
-import { IWorkspaceService } from '../../workspace/common/workspaceService';
-import { IAdoCodeSearchService } from '../common/adoCodeSearchService';
-import { IGithubCodeSearchService } from '../common/githubCodeSearchService';
-import { RemoteCodeSearchError, RemoteCodeSearchIndexState, RemoteCodeSearchIndexStatus } from '../common/remoteCodeSearch';
-import { ICodeSearchAuthenticationService } from './codeSearchRepoAuth';
+import { Result } from '../../../../util/common/result';
+import { TelemetryCorrelationId } from '../../../../util/common/telemetryCorrelationId';
+import { DeferredPromise, IntervalTimer, raceCancellationError } from '../../../../util/vs/base/common/async';
+import { CancellationToken, CancellationTokenSource } from '../../../../util/vs/base/common/cancellation';
+import { isCancellationError } from '../../../../util/vs/base/common/errors';
+import { Emitter, Event } from '../../../../util/vs/base/common/event';
+import { Disposable } from '../../../../util/vs/base/common/lifecycle';
+import { ResourceMap } from '../../../../util/vs/base/common/map';
+import { IInstantiationService } from '../../../../util/vs/platform/instantiation/common/instantiation';
+import { IAuthenticationService } from '../../../authentication/common/authentication';
+import { AdoRepoId, GithubRepoId, IGitService, ResolvedRepoRemoteInfo } from '../../../git/common/gitService';
+import { Change } from '../../../git/vscode/git';
+import { LogExecTime, logExecTime } from '../../../log/common/logExecTime';
+import { ILogService } from '../../../log/common/logService';
+import { IAdoCodeSearchService } from '../../../remoteCodeSearch/common/adoCodeSearchService';
+import { IGithubCodeSearchService } from '../../../remoteCodeSearch/common/githubCodeSearchService';
+import { RemoteCodeSearchError, RemoteCodeSearchIndexState, RemoteCodeSearchIndexStatus } from '../../../remoteCodeSearch/common/remoteCodeSearch';
+import { ICodeSearchAuthenticationService } from '../../../remoteCodeSearch/node/codeSearchRepoAuth';
+import { isGitHubRemoteRepository } from '../../../remoteRepositories/common/utils';
+import { CodeSearchRepoTracker, RepoInfo, TrackedRepoState, TrackedRepoStatus } from './repoTracker';
 
 export enum RepoStatus {
 	/** We could not resolve this repo */
@@ -70,14 +62,8 @@ export enum RepoStatus {
 	BuildingIndex = 'BuildingIndex',
 
 	/** The remote index is ready and usable */
-	Ready = 'Ready',
+	Ready = 'Ready'
 }
-
-export interface RepoInfo {
-	readonly rootUri: URI;
-}
-
-
 export interface ResolvedRepoEntry {
 	readonly status: RepoStatus.NotYetIndexed | RepoStatus.NotIndexable | RepoStatus.BuildingIndex | RepoStatus.CouldNotCheckIndexStatus | RepoStatus.NotAuthorized;
 	readonly repo: RepoInfo;
@@ -108,10 +94,9 @@ export type RepoEntry =
 		readonly repo: RepoInfo;
 		readonly remoteInfo: ResolvedRepoRemoteInfo;
 		readonly initTask: InitTask;
-	}
-	| ResolvedRepoEntry
-	| IndexedRepoEntry
-	;
+	} |
+	ResolvedRepoEntry |
+	IndexedRepoEntry;
 
 export type BuildIndexTriggerReason = 'auto' | 'manual';
 
@@ -169,45 +154,6 @@ export interface CodeSearchDiff {
 	readonly mayBeOutdated?: boolean;
 }
 
-/**
- * Ids used to identify the type of remote for telemetry purposes
- *
- * Do not change these values as they are used in telemetry.
- */
-enum GitRemoteTypeForTelemetry {
-	NoRemotes = 0,
-	Unknown = 1,
-
-	Github = 2,
-	Ghe = 3,
-
-	// Unsupported
-	AzureDevOps = 4,
-	VisualStudioDotCom = 5,
-	GitLab = 6,
-	BitBucket = 7,
-}
-
-const remoteHostTelemetryIdMapping = new Map<string, GitRemoteTypeForTelemetry>([
-	['github.com', GitRemoteTypeForTelemetry.Github],
-	['ghe.com', GitRemoteTypeForTelemetry.Ghe],
-
-	['dev.azure.com', GitRemoteTypeForTelemetry.AzureDevOps],
-	['visualstudio.com', GitRemoteTypeForTelemetry.VisualStudioDotCom],
-	['gitlab.com', GitRemoteTypeForTelemetry.GitLab],
-	['bitbucket.org', GitRemoteTypeForTelemetry.BitBucket],
-]);
-
-function getRemoteTypeForTelemetry(remoteHost: string): GitRemoteTypeForTelemetry {
-	remoteHost = remoteHost.toLowerCase();
-	for (const [key, value] of remoteHostTelemetryIdMapping) {
-		if (remoteHost === key || remoteHost.endsWith('.' + key)) {
-			return value;
-		}
-	}
-
-	return GitRemoteTypeForTelemetry.Unknown;
-}
 
 /**
  * Tracks all repositories in the workspace that have been indexed for code search.
@@ -244,7 +190,7 @@ export class CodeSearchRepoManager extends Disposable {
 		@ICodeSearchAuthenticationService private readonly _codeSearchAuthService: ICodeSearchAuthenticationService,
 		@IGithubCodeSearchService private readonly _githubCodeSearchService: IGithubCodeSearchService,
 		@IGitService private readonly _gitService: IGitService,
-		@ILogService private readonly _logService: ILogService,
+		@ILogService private readonly _logService: ILogService
 	) {
 		super();
 
@@ -261,6 +207,7 @@ export class CodeSearchRepoManager extends Disposable {
 		const refreshInterval = this._register(new IntervalTimer());
 		refreshInterval.cancelAndSet(() => this.updateIndexedCommitForAllRepos(), 5 * 60 * 1000); // 5 minutes
 
+
 		// When the authentication state changes, update repos
 		this._register(Event.any(
 			this._authenticationService.onDidAuthenticationChange,
@@ -270,7 +217,7 @@ export class CodeSearchRepoManager extends Disposable {
 		}));
 
 		this._register(Event.any(
-			this._authenticationService.onDidAdoAuthenticationChange,
+			this._authenticationService.onDidAdoAuthenticationChange
 		)(() => {
 			this.updateRepoStatuses('ado');
 		}));
@@ -292,7 +239,7 @@ export class CodeSearchRepoManager extends Disposable {
 
 					// And make sure they have done their initial checks.
 					// After this the repos may still be left polling github but we've done at least one check
-					await Promise.all(Array.from(this._repos.values(), async repo => {
+					await Promise.all(Array.from(this._repos.values(), async (repo) => {
 						if (repo.status === RepoStatus.CheckingStatus) {
 							try {
 								await repo.initTask.p;
@@ -342,18 +289,19 @@ export class CodeSearchRepoManager extends Disposable {
 		return this._repos.get(repo.repo.rootUri)?.status ?? repo.status;
 	}
 
-	private addOrUpdateTrackedRepo(info: TrackedRepoInfo) {
+	private addOrUpdateTrackedRepo(info: TrackedRepoState) {
 		switch (info.status) {
-			case TrackedRepoStatus.NotResolvable: {
-				this.updateRepoEntry(info.repo, { status: RepoStatus.NotResolvable, repo: info.repo });
-				return;
-			}
 			case TrackedRepoStatus.Resolving: {
 				this.updateRepoEntry(info.repo, { status: RepoStatus.Resolving, repo: info.repo });
 				return;
 			}
 			case TrackedRepoStatus.Resolved: {
-				return this.openGitRepo(info.repo, info.remoteInfo);
+				if (info.resolvedRemoteInfo) {
+					return this.openGitRepo(info.repo, info.resolvedRemoteInfo);
+				} else {
+					// We found a git repo but not one a type we know about
+					this.updateRepoEntry(info.repo, { status: RepoStatus.NotResolvable, repo: info.repo });
+				}
 			}
 		}
 	}
@@ -494,8 +442,7 @@ export class CodeSearchRepoManager extends Disposable {
 		this._logService.trace(`RepoTracker.TriggerRemoteIndexing(${triggerReason}).Repos: ${JSON.stringify(Array.from(this._repos.values(), r => ({
 			rootUri: r.repo.rootUri.toString(),
 			status: r.status,
-		})), null, 4)
-			} `);
+		})), null, 4)} `);
 
 		const allRepos = Array.from(this._repos.values());
 		if (!allRepos.length) {
@@ -625,7 +572,6 @@ export class CodeSearchRepoManager extends Disposable {
 
 		// TODO: only handles first repos of each type, but our other services also don't track tokens for multiple
 		// repos in a workspace right now
-
 		const firstGithubRepo = notAuthorizedRepos.find(repo => repo.remoteInfo.repoId.type === 'github');
 		if (firstGithubRepo) {
 			await this._codeSearchAuthService.tryAuthenticating(firstGithubRepo);
@@ -781,388 +727,6 @@ export class CodeSearchRepoManager extends Disposable {
 					() => {
 						// Noop
 					});
-		}
-	}
-}
-
-enum TrackedRepoStatus {
-	Resolving = 'Resolving',
-	NotResolvable = 'NotResolvable',
-	Resolved = 'Resolved',
-}
-
-type TrackedRepoInfo =
-	{
-		readonly status: TrackedRepoStatus.NotResolvable;
-		readonly repo: RepoInfo;
-	} | {
-		readonly status: TrackedRepoStatus.Resolving;
-		readonly repo: RepoInfo;
-		readonly initTask: InitTask;
-	} | {
-		readonly status: TrackedRepoStatus.Resolved;
-		readonly repo: RepoInfo;
-		readonly remoteInfo: ResolvedRepoRemoteInfo;
-	}
-	;
-
-class CodeSearchRepoTracker extends Disposable {
-	private readonly _repos = new ResourceMap<TrackedRepoInfo>();
-
-	private readonly _onDidAddOrUpdateRepo = this._register(new Emitter<TrackedRepoInfo>());
-	public readonly onDidAddOrUpdateRepo = this._onDidAddOrUpdateRepo.event;
-
-	private readonly _onDidRemoveRepo = this._register(new Emitter<TrackedRepoInfo>());
-	public readonly onDidRemoveRepo = this._onDidRemoveRepo.event;
-
-	private readonly _initializedGitReposP: CancelablePromise<void>;
-	private readonly _initializedGitHubRemoteReposP: CancelablePromise<void>;
-
-	private _isDisposed = false;
-
-	constructor(
-		@IGitExtensionService private readonly _gitExtensionService: IGitExtensionService,
-		@IGitService private readonly _gitService: IGitService,
-		@ILogService private readonly _logService: ILogService,
-		@ISimulationTestContext private readonly _simulationTestContext: ISimulationTestContext,
-		@ITelemetryService private readonly _telemetryService: ITelemetryService,
-		@IWorkspaceService private readonly _workspaceService: IWorkspaceService,
-	) {
-		super();
-
-		this._initializedGitReposP = createCancelablePromise(async (token) => {
-			this._logService.trace(`CodeSearchRepoTracker.tryInitGitRepos(): started`);
-
-			try {
-				if (!this._gitService.isInitialized) {
-					this._logService.trace(`CodeSearchRepoTracker.tryInitGitRepos(): Git service not initialized. Waiting for init signal.`);
-					const finishInitTimeout = 30_000;
-					await raceCancellationError(raceTimeout(new Promise<void>(resolve => this._gitService.onDidFinishInitialization(() => resolve())), finishInitTimeout), token);
-					if (this._isDisposed) {
-						return;
-					}
-				}
-
-				this._logService.trace(`CodeSearchRepoTracker.tryInitGitRepos(): Found initial repos: [${this._gitService.repositories.map(repo => repo.rootUri.toString())}].`);
-
-				const openPromises = this._gitService.repositories.map(repo => this.openGitRepo(repo));
-
-				this._register(this._gitService.onDidOpenRepository(repo => this.openGitRepo(repo)));
-				this._register(this._gitService.onDidCloseRepository(repo => this.closeRepo(repo)));
-
-				await raceCancellationError(Promise.allSettled(openPromises), token);
-				this._logService.trace(`CodeSearchRepoTracker.tryInitGitRepos(): Complete`);
-			} catch (e) {
-				this._logService.error(`CodeSearchRepoTracker.tryInitGitRepos(): Error occurred during initialization: ${e}`);
-			}
-		});
-
-		this._initializedGitHubRemoteReposP = createCancelablePromise(async (token) => {
-			try {
-				const githubRemoteRepos = this._workspaceService.getWorkspaceFolders().filter(isGitHubRemoteRepository);
-				if (!githubRemoteRepos.length) {
-					return;
-				}
-
-				this._logService.trace(`CodeSearchRepoTracker.initGithubRemoteRepos(): started`);
-
-				await raceCancellationError(
-					Promise.all(githubRemoteRepos.map(workspaceRoot => {
-						const githubRepoIdParts = workspaceRoot.path.slice(1).split('/');
-						return this.openGithubRemoteRepo(workspaceRoot, new GithubRepoId(githubRepoIdParts[0], githubRepoIdParts[1]));
-					})),
-					token);
-				this._logService.trace(`CodeSearchRepoTracker.initGithubRemoteRepos(): complete`);
-			} catch (e) {
-				this._logService.error(`CodeSearchRepoTracker.initGithubRemoteRepos(): Error occurred during initialization: ${e}`);
-			}
-		});
-	}
-
-	public override dispose(): void {
-		super.dispose();
-
-		this._isDisposed = true;
-
-		this._initializedGitReposP.cancel();
-		this._initializedGitHubRemoteReposP.cancel();
-
-		for (const repo of this._repos.values()) {
-			if (repo.status === TrackedRepoStatus.Resolving) {
-				repo.initTask.cts.cancel();
-			}
-		}
-		this._repos.clear();
-	}
-
-	async initialize(): Promise<void> {
-		try {
-			// Wait for the initial repos to be discovered
-			await Promise.all([
-				this._initializedGitReposP,
-				this._initializedGitHubRemoteReposP
-			]);
-
-			// And wait for each repo to resolve
-			await Promise.all(Array.from(this._repos.values(), async repo => {
-				if (repo.status === TrackedRepoStatus.Resolving) {
-					try {
-						await repo.initTask.p;
-					} catch (error) {
-						this._logService.error(`Error during repo initialization: ${error}`);
-					}
-				}
-			}));
-		} catch (e) {
-			// Noop
-		}
-	}
-
-	private updateRepoEntry(repo: RepoInfo, entry: TrackedRepoInfo): void {
-		this._repos.set(repo.rootUri, entry);
-		this._onDidAddOrUpdateRepo.fire(entry);
-	}
-
-	@LogExecTime(self => self._logService, 'CodeSearchRepoTracker::openGitRepo')
-	private async openGitRepo(repo: RepoContext): Promise<void> {
-		this._logService.trace(`CodeSearchRepoTracker.openGitRepo(${repo.rootUri})`);
-
-		const existing = this._repos.get(repo.rootUri);
-		if (existing) {
-			if (existing.status === TrackedRepoStatus.Resolving) {
-				try {
-					return await existing.initTask.p;
-				} catch (e) {
-					if (isCancellationError(e)) {
-						return;
-					}
-
-					throw e;
-				}
-			}
-		}
-
-		const initDeferredP = new DeferredPromise<void>();
-		const initTask: InitTask = {
-			p: initDeferredP.p,
-			cts: new CancellationTokenSource(),
-		};
-
-		const initToken = initTask.cts.token;
-		(async () => {
-			try {
-				// Do a status check to make sure the repo info is fully loaded
-				// See #12954
-				await this._gitExtensionService.getExtensionApi()?.getRepository(repo.rootUri)?.status();
-			} catch {
-				this._logService.trace(`CodeSearchRepoTracker.openRepo(${repo.rootUri}). git status check failed.`);
-				// Noop, may still be ok even if the status check failed
-			}
-
-			if (initToken.isCancellationRequested) {
-				return;
-			}
-
-			const updatedRepo = await this._gitService.getRepository(repo.rootUri);
-			if (!updatedRepo && !this._simulationTestContext.isInSimulationTests) {
-				this._logService.trace(`CodeSearchRepoTracker.openRepo(${repo.rootUri}). No current repo found after status check.`);
-
-				/* __GDPR__
-					"codeSearchRepoTracker.openGitRepo.error.noCurrentRepo" : {
-						"owner": "mjbvz",
-						"comment": "Information about errors when trying to resolve a remote"
-					}
-				*/
-				this._telemetryService.sendMSFTTelemetryEvent('codeSearchRepoTracker.openGitRepo.error.noCurrentRepo');
-
-				this.updateRepoEntry(repo, { status: TrackedRepoStatus.NotResolvable, repo });
-				return;
-			}
-
-			if (updatedRepo) {
-				repo = updatedRepo;
-			}
-			this._repos.set(repo.rootUri, { status: TrackedRepoStatus.Resolving, repo, initTask });
-
-			const remoteInfos = await this.getRemoteInfosForRepo(repo);
-			if (initToken.isCancellationRequested) {
-				return;
-			}
-
-			let remoteTelemetryType: GitRemoteTypeForTelemetry;
-			if (remoteInfos.length) {
-				const primaryRemote = remoteInfos[0];
-				const remoteHost = primaryRemote.fetchUrl ? parseRemoteUrl(primaryRemote.fetchUrl) : undefined;
-				remoteTelemetryType = remoteHost ? getRemoteTypeForTelemetry(remoteHost.host) : GitRemoteTypeForTelemetry.Unknown;
-			} else {
-				const allRemotes = Array.from(getOrderedRemoteUrlsFromContext(repo));
-				if (allRemotes.length === 0) {
-					remoteTelemetryType = GitRemoteTypeForTelemetry.NoRemotes;
-				} else {
-					for (const remote of allRemotes) {
-						if (remote) {
-							const remoteHost = parseRemoteUrl(remote);
-							if (remoteHost) {
-								const telemetryId = getRemoteTypeForTelemetry(remoteHost.host);
-								if (telemetryId !== GitRemoteTypeForTelemetry.Unknown) {
-									remoteTelemetryType = telemetryId;
-									break;
-								}
-							}
-						}
-					}
-				}
-				remoteTelemetryType ??= GitRemoteTypeForTelemetry.Unknown;
-			}
-			/* __GDPR__
-				"codeSearchRepoTracker.openGitRepo.remoteInfo" : {
-					"owner": "mjbvz",
-					"comment": "Information about the remote",
-					"resolvedRemoteType": { "classification": "SystemMetaData", "purpose": "FeatureInsight", "isMeasurement": true, "comment": "Identifies the primary remote's type " }
-				}
-			*/
-			this._telemetryService.sendMSFTTelemetryEvent('codeSearchRepoTracker.openGitRepo.remoteInfo', {}, {
-				resolvedRemoteType: remoteTelemetryType,
-			});
-
-			if (!remoteInfos.length) {
-				this._logService.trace(`CodeSearchRepoTracker.openRepo(${repo.rootUri}). No valid github remote found. Remote urls: ${JSON.stringify(Array.from(getOrderedRemoteUrlsFromContext(repo)))}.`);
-
-				this._telemetryService.sendInternalMSFTTelemetryEvent('codeSearchRepoTracker.error.couldNotResolveRemote.internal', {
-					remoteUrls: JSON.stringify(coalesce(repo.remoteFetchUrls ?? [])),
-				});
-
-				/* __GDPR__
-					"codeSearchRepoTracker.openGitRepo.error.couldNotResolveRemote" : {
-						"owner": "mjbvz",
-						"comment": "Information about errors when trying to resolve a remote",
-						"repoRemoteFetchUrlsCount": { "classification": "SystemMetaData", "purpose": "FeatureInsight", "isMeasurement": true, "comment": "Number of remote fetch urls on the git repo" }
-					}
-				*/
-				this._telemetryService.sendMSFTTelemetryEvent('codeSearchRepoTracker.openGitRepo.error.couldNotResolveRemote', {}, {
-					repoRemoteFetchUrlsCount: repo.remoteFetchUrls?.length ?? 0,
-				});
-
-				this.updateRepoEntry(repo, { status: TrackedRepoStatus.NotResolvable, repo });
-				return;
-			}
-
-			// TODO: Support checking index status for multiple remotes
-			const primaryRemote = remoteInfos[0];
-			this.updateRepoEntry(repo, {
-				status: TrackedRepoStatus.Resolved,
-				repo,
-				remoteInfo: primaryRemote,
-			});
-		})()
-			.catch(() => { })
-			.finally(() => {
-				initDeferredP.complete();
-			});
-
-		this._repos.set(repo.rootUri, {
-			status: TrackedRepoStatus.Resolving,
-			repo,
-			initTask
-		});
-	}
-
-	private async openGithubRemoteRepo(rootUri: URI, githubId: GithubRepoId): Promise<void> {
-		this._logService.trace(`CodeSearchRepoTracker.openGithubRemoteRepo(${rootUri})`);
-
-		const existing = this._repos.get(rootUri);
-		if (existing) {
-			if (existing.status === TrackedRepoStatus.Resolving) {
-				return existing.initTask.p;
-			}
-		}
-
-		this._repos.set(rootUri, {
-			status: TrackedRepoStatus.Resolved,
-			repo: { rootUri },
-			remoteInfo: {
-				repoId: githubId,
-				fetchUrl: undefined,
-			}
-		});
-	}
-
-	private closeRepo(repo: RepoContext) {
-		this._logService.trace(`CodeSearchRepoTracker.closeRepo(${repo.rootUri})`);
-
-		const repoEntry = this._repos.get(repo.rootUri);
-		if (!repoEntry) {
-			return;
-		}
-
-		if (repoEntry.status === TrackedRepoStatus.Resolving) {
-			repoEntry.initTask.cts.cancel();
-		}
-
-		this._onDidRemoveRepo.fire(repoEntry);
-		this._repos.delete(repo.rootUri);
-	}
-
-	private async getRemoteInfosForRepo(repo: RepoContext): Promise<ResolvedRepoRemoteInfo[]> {
-		const remoteInfos = Array.from(getOrderedRepoInfosFromContext(repo));
-
-		// Fallback to checking the SSH config if no remotes were found
-		if (!remoteInfos.length) {
-			const other = await this.getGithubRemoteFromSshConfig(repo);
-			if (other) {
-				remoteInfos.push(other);
-			}
-		}
-
-		// For now always prefer the github remotes
-		remoteInfos.sort((a, b) => {
-			if (a.repoId.type === 'github' && b.repoId.type !== 'github') {
-				return -1;
-			} else if (b.repoId.type === 'github' && a.repoId.type !== 'github') {
-				return 1;
-			}
-			return 0;
-		});
-
-		return remoteInfos;
-	}
-
-	private async getGithubRemoteFromSshConfig(repo: RepoContext): Promise<ResolvedRepoRemoteInfo | undefined> {
-		if (repo.rootUri.scheme !== Schemas.file) {
-			return;
-		}
-
-		try {
-			const execAsync = promisify(exec);
-			const { stdout, stderr } = await execAsync('git -c credential.interactive=never fetch --dry-run', {
-				cwd: repo.rootUri.fsPath,
-				env: {
-					GIT_SSH_COMMAND: 'ssh -v -o BatchMode=yes'
-				}
-			});
-
-			const output = stdout + '\n' + stderr;
-
-			const authMatch = output.match(/^Authenticated to ([^\s]+)\s/m);
-			const fromMatch = output.match(/^From ([^:]+):([^/]+)\/([^\s]+)$/m);
-
-			if (authMatch && fromMatch) {
-				const authenticatedTo = authMatch[1];
-				const owner = fromMatch[2];
-				const repo = fromMatch[3].replace(/\.git$/, '');
-				const remoteUrl = `ssh://${authenticatedTo}/${owner}/${repo}`;
-
-				const githubRepoId = getGithubRepoIdFromFetchUrl(remoteUrl);
-				if (githubRepoId) {
-					return {
-						repoId: githubRepoId,
-						fetchUrl: remoteUrl
-					};
-				}
-			}
-			return undefined;
-		} catch (e) {
-			return undefined;
 		}
 	}
 }
